Removed PetContext because it was unneeeded
